Remove patching of `reveal_type()` and `reveal_locals()`
This is no longer necessary as we have enabled the mypy error code to
prevent this: `unimported_reveal`. Now any stray uses of these functions
will be caught in CI and not make it into a release of django-stubs.

Keep the `include_builtins` argument in case anyone is using it, but
mark it as deprecated.